pyproject.toml
[project]
name = "printer-music"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "mido>=1.3.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.13",
    "types-pyserial>=3.5.0.20251001",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "mido.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py312"
